+++ PhotoCollection/Sources/App/AppDelegate.swift
@@ -42,5 +42,4 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         
         return true
     }
-    
 }
