+++ PhotoCollection/Sources/Model/Model.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 struct Item {
-    
     let icon: UIImage?
     let title: String
     let detailTextLabel: String?
@@ -29,7 +28,7 @@ struct Section {
 
 extension Section {
     static func getData() -> [Section] {
-        return [
+        [
             Section(type: .twoRollHorizontalView, title: "Мои альбомы", options: [
                 Item(icon: UIImage(named: "image0"), title: "Недавние", detailTextLabel: "2 432"),
                 Item(icon: UIImage(named: "image1"), title: "Избранное", detailTextLabel: "100"),
@@ -38,12 +37,14 @@ extension Section {
                 Item(icon: UIImage(named: "image4"), title: "WhatsApp", detailTextLabel: "45"),
                 Item(icon: UIImage(named: "image5"), title: "Dropbox", detailTextLabel: "4")
             ]),
+            
             Section(type: .oneRollHorizontalView, title: "Общие места", options: [
                 Item(icon: UIImage(named: "image6"), title: "Недавние", detailTextLabel: "2 432"),
                 Item(icon: UIImage(named: "image7"), title: "Instagram", detailTextLabel: "104"),
                 Item(icon: UIImage(named: "image8"), title: "Travel", detailTextLabel: "432"),
                 Item(icon: UIImage(named: "image9"), title: "Избранное", detailTextLabel: "215")
             ]),
+            
             Section(type: .verticalView, title: "Типы медиафайлов", options: [
                 Item(icon: UIImage(named: "Видео"), title: "Видео", detailTextLabel: "432"),
                 Item(icon: UIImage(named: "Селфи"), title: "Селфи", detailTextLabel: "57"),
