+++ PhotoCollection/Sources/Scenes/FirstTab/FirstTabViewController.swift
@@ -1,15 +1,16 @@
 //
-//  SecondTabViewController.swift
+//  FirstTabViewController.swift
 //  PhotoCollection
 //
 //  Created by Artyom Guzenko on 27.04.2022.
 //
+
 import UIKit
 
-class SecondTabViewController: UIViewController {
+final class FirstTabViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
-        title = "Для Вас"
+        title = "Медиатека"
         navigationController?.navigationBar.prefersLargeTitles = true
         view.backgroundColor = .systemGray6
     }
