+++ PhotoCollection/Sources/Scenes/SecondTab/SecondTabViewController.swift
@@ -1,5 +1,5 @@
 //
-//  FirstTabViewController.swift
+//  SecondTabViewController.swift
 //  PhotoCollection
 //
 //  Created by Artyom Guzenko on 27.04.2022.
@@ -7,11 +7,10 @@
 
 import UIKit
 
-class FirstTabViewController: UIViewController {
-    
+final class SecondTabViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
-        title = "Медиатека"
+        title = "Для Вас"
         navigationController?.navigationBar.prefersLargeTitles = true
         view.backgroundColor = .systemGray6
     }
