--- PhotoCollection/Sources/Scenes/SectionHeader.swift
@@ -1,66 +1,0 @@
-//
-//  SectionHeader.swift
-//  PhotoCollection
-//
-//  Created by Artyom Guzenko on 27.04.2022.
-//
-
-import UIKit
-
-class SectionHeader: UICollectionReusableView {
-    
-    static let reuseId = "SectionHeader"
-    
-    let title = UILabel()
-    let line = UILabel()
-     
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        viewHierarchy()
-        setupLayout()
-        setupElements()
-    }
-    
-    private func viewHierarchy() {
-        addSubview(line)
-        addSubview(title)
-    }
-    
-    private func setupLayout() {
-        line.translatesAutoresizingMaskIntoConstraints = false
-        line.leadingAnchor.constraint(equalTo: leadingAnchor, constant: SectionHeaderMetric.lineLeadingArchorConstant).isActive = true
-        line.trailingAnchor.constraint(equalTo: trailingAnchor).isActive = true
-        line.topAnchor.constraint(equalTo: topAnchor).isActive = true
-        line.heightAnchor.constraint(equalToConstant: SectionHeaderMetric.lineHeightAnchorConstant).isActive = true
-
-
-        title.translatesAutoresizingMaskIntoConstraints = false
-        title.leadingAnchor.constraint(equalTo: leadingAnchor, constant: SectionHeaderMetric.titleLeadingArchorConstant).isActive = true
-        title.centerYAnchor.constraint(equalTo: centerYAnchor, constant: SectionHeaderMetric.titleCenterYAnchorConstant).isActive = true
-        title.heightAnchor.constraint(equalToConstant: SectionHeaderMetric.titleHeightAnchorConstant).isActive = true
-    }
-    
-    private func setupElements() {
-        title.textColor = .black
-        title.font = UIFont.boldSystemFont(ofSize: SectionHeaderMetric.titleFontSize)
-        
-        line.backgroundColor = UIColor.systemGray3
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-// MARK: - Constants
-
-enum SectionHeaderMetric {
-    static let lineLeadingArchorConstant = 10.0
-    static let lineHeightAnchorConstant = 1.0
-    
-    static let titleLeadingArchorConstant = 10.0
-    static let titleCenterYAnchorConstant = 20.0
-    static let titleHeightAnchorConstant = 20.0
-    static let titleFontSize = 25.0
-}
-
